backend/requirements.txt
# Flask and extensions
Flask==2.3.2
Flask-CORS==4.0.0         # For frontend-backend communication

# Database
Flask-SQLAlchemy==3.0.3   # ORM for PostgreSQL
Flask-Migrate==4.0.4      # Database migrations
psycopg2-binary==2.9.6    # PostgreSQL adapter

# api development 
python-dotenv==1.0.0      # Environment variables
requests==2.31.0          # For fetching external API data
finnhub-python==2.4.0     # Finnhub API client

# optional packages 
pandas==2.0.3             # If analyzing trade data
python-dateutil==2.8.2    # For date handling

# WebSocket support
websocket-client==1.6.2   # For real-time data streaming
